Consolidate wholeprog.go into deps.go and rename to Main prefix
- Move all whole-program SSA analysis code from wholeprog.go into deps.go
- Rename CreateWholeProgDeps to CreateMainDeps
- Rename analyzeWholeProgFuncDeps to analyzeMainFuncDeps
- Restrict Test* root detection to functions defined in _test.go files
